First version of serperated track rendering

+++ src/Ringbuffer.cpp
@@ -9,7 +9,7 @@ using namespace agbplay;
  * public Ringbuffer
  */
 
-Ringbuffer::Ringbuffer(uint32_t elementCount)
+Ringbuffer::Ringbuffer(size_t elementCount)
     : bufData(elementCount)
 {
     freePos = 0;
@@ -22,20 +22,20 @@ Ringbuffer::~Ringbuffer()
 {
 }
 
-void Ringbuffer::Put(float *inData, uint32_t nElements)
+void Ringbuffer::Put(float *inData, size_t nElements)
 {
     boost::mutex::scoped_lock lock(countLock);
     while (freeCount < nElements){
         sig.wait(lock);
     }
     while (nElements > 0) {
-        uint32_t count = put(inData, nElements);
+        size_t count = put(inData, nElements);
         inData += count;
         nElements -= count;
     }
 }
 
-void Ringbuffer::Take(float *outData, uint32_t nElements)
+void Ringbuffer::Take(float *outData, size_t nElements)
 {
     if (dataCount < nElements) {
         // underrun
@@ -44,7 +44,7 @@ void Ringbuffer::Take(float *outData, uint32_t nElements)
         // output
         boost::mutex::scoped_lock lock(countLock);
         while (nElements > 0) {
-            uint32_t count = take(outData, nElements);
+            size_t count = take(outData, nElements);
             outData += count;
             nElements -= count;
         }
@@ -62,13 +62,13 @@ void Ringbuffer::Clear()
  * private Ringbuffer
  */
 
-uint32_t Ringbuffer::put(float *inData, uint32_t nElements)
+size_t Ringbuffer::put(float *inData, size_t nElements)
 {
     bool wrap = nElements > bufData.size() - freePos;
-    uint32_t count;
-    uint32_t newfree;
+    size_t count;
+    size_t newfree;
     if (wrap) {
-        count = uint32_t(bufData.size() - freePos);
+        count = size_t(bufData.size() - freePos);
         newfree = 0;
     } else {
         count = nElements;
@@ -81,13 +81,13 @@ uint32_t Ringbuffer::put(float *inData, uint32_t nElements)
     return count;
 }
 
-uint32_t Ringbuffer::take(float *outData, uint32_t nElements)
+size_t Ringbuffer::take(float *outData, size_t nElements)
 {
     bool wrap = nElements > bufData.size() - dataPos;
-    uint32_t count;
-    uint32_t newdata;
+    size_t count;
+    size_t newdata;
     if (wrap) {
-        count = uint32_t(bufData.size() - dataPos);
+        count = size_t(bufData.size() - dataPos);
         newdata = 0;
     } else {
         count = nElements;
